I was returning content with 204 messages which I SHOULD NOT. Fixed that. I edited the content negotiation to change the order of the expect header if a call is made form Javascript. Added Location headers for all 201 messages.
The test files now test for the location header.
If the test is run from within the server itself, the 204 errors are not seen, but the same tests fail if they are run from another lcoation.

Added docs detailing the 204 error and also the expect headers received from different browsers.